Refactor LanguageCodes into shared common module
