+++ packages/common/index.ts
@@ -16,3 +16,4 @@ export * from './src/EmailHelper';
 export * from './src/EventHelper';
 export * from './src/form/FormHelperCommon';
 export * from './src/RatingHelper';
+export * from './src/LanguageCodes';
